requirements.txt
streamlit==1.31.1
qdrant-client==1.7.3
sentence-transformers==2.6.1
pypdf==4.0.2
python-docx==1.1.0
python-dotenv==1.0.1
requests==2.31.0
tqdm==4.66.1
numpy==1.26.4